Final state of weekend assignment.
No idea what happened, got confused as to reason for infinite server requests (unintentional recursion in the DBManager class) but during that fix tried a number of different ways to restructure the interfaces. Then was unable to revert changes to previous state and eventually tried to replicate with a fresh build ~3am but gave up after 4:30.

Somehow during this files were removed and have been unable to recover. Will post the latest state of the redux as well.